requirements.txt
websockets
kafka-python
python-dotenv